Primeiro commit - Criando o projeto básico e  realizando dois testes básicos

+++ Demo/Claculadora.cs
@@ -8,12 +8,12 @@ namespace Demo
 {
     public class Claculadora
     {
-        public int Somar(int a, int b)
+        public double Somar(double a, double b)
         {
             return a + b;
         }
 
-        public double Dividir(int a, int b)
+        public double Dividir(double a, double b)
         {
             return a / b;
         }
